auth-service/requirements.txt
fastapi
uvicorn
pymongo
python-jose
passlib
cryptography